backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.2

# Pydantic settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Gemini AI
google-generativeai==0.3.1

# Geospatial
shapely==2.0.2
geopy==2.4.1

# Utilities
python-dotenv==1.0.0
